Move get-loader-options.js
